requirements.txt
openai
pinecone
tiktoken
python-dotenv
numpy
scikit-learn
pyyaml
cohere
python-dotenv
tqdm
pypdf
langchain   
langchain-text-splitters